+++ modules/carto/src/api/index.ts
@@ -1,5 +1,3 @@
 export {CartoAPIError} from './carto-api-error';
-export type {APIErrorContext} from './carto-api-error';
-export {FORMATS, TILE_FORMATS, MAP_TYPES, API_VERSIONS} from './maps-api-common';
-export type {QueryParameters} from './maps-api-common';
 export {fetchMap} from './maps-v3-client';
+export type {APIErrorContext, Format, MapType, RequestType, QueryParameters} from './types';
