+++ modules/carto/src/index.ts
@@ -23,8 +23,8 @@ export {default as BASEMAP} from './basemap';
 export {default as colorBins} from './style/color-bins-style';
 export {default as colorCategories} from './style/color-categories-style';
 export {default as colorContinuous} from './style/color-continuous-style';
-export {FORMATS, TILE_FORMATS, MAP_TYPES, API_VERSIONS, CartoAPIError, fetchMap} from './api';
-export type {APIErrorContext, QueryParameters} from './api';
+export {CartoAPIError, fetchMap} from './api';
+export type {APIErrorContext, Format, MapType, RequestType, QueryParameters} from './api';
 
 import {
   h3QuerySource,
