+++ modules/carto/src/sources/utils.ts
@@ -1,7 +1,14 @@
-import {APIErrorContext, CartoAPIError} from '../api/carto-api-error';
-import {encodeParameter, MapType} from '../api/maps-api-common';
+import {CartoAPIError} from '../api/carto-api-error';
 import {DEFAULT_HEADERS, DEFAULT_PARAMETERS, MAX_GET_LENGTH} from './common';
 import {buildMapsUrlFromBase} from '../config';
+import type {APIErrorContext, MapType} from '../api/types';
+
+/**
+ * Simple encode parameter
+ */
+function encodeParameter(name: string, value: string | boolean | number): string {
+  return `${name}=${encodeURIComponent(value)}`;
+}
 
 const REQUEST_CACHE = new Map();
 export async function requestWithParameters<T = any>({
